# lusevokteren-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FjordVind Lusevokteren API
  description: |
    REST API for sea lice monitoring in Norwegian aquaculture.

    ## Authentication
    Most endpoints require authentication via JWT Bearer token.

    ```
    Authorization: Bearer <your-token>
    ```

    ## Demo Mode
    In development, demo tokens are available:
    - `demo_token_admin` - Admin access
    - `demo_token_driftsleder` - Operations manager
    - `demo_token_røkter` - Fish farmer

  version: 1.1.0
  contact:
    name: FjordVind Support
    email: support@fjordvind.no
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.fjordvind.no
    description: Production server
...
paths:
...
  /api/mortality:
    get:
      summary: List mortality records
      tags: [Mortality]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of mortality records

    post:
      summary: Create mortality record
      tags: [Mortality]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMortalityRequest'
      responses:
        '201':
          description: Mortality record created

  # ========== ENVIRONMENT ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateMortalityRequest:
      type: object
      required:
        - merd_id
        - record_date
      properties:
        merd_id:
          type: string
          format: uuid
        record_date:
          type: string
          format: date
        salmon_dead:
          type: integer
        salmon_cause:
          type: string
        cleaner_fish_dead:
          type: integer
        mortality_category:
          type: string
          enum: [normal, behandling, sykdom, håndtering, predator, miljø, annet]
...